# docs/openapi/openapi.v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKDY RAG Server API",
    "description": "HTTP API for the SKDY RAG Server. Mounted under `/api/v1`.\n\n**Contract**: the OpenAPI schema is the single source of truth for the Web UI; client types are generated from it.\n\n**Error envelope**: every 4xx/5xx returns `{\"error\": {\"code\": \"...\", \"message\": \"...\", \"request_id\": \"...\", \"details\": {}}}`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List all documents across collections",
        "description": "Server-side paginated listing of documents across **every** collection.\n\nMirrors the per-collection paging path: the page is selected in\nSQLite (newest-first across the whole corpus) and only the current\npage's chunk / image counts are queried — no full-corpus materialisation.",
        "operationId": "list_all_documents_api_v1_documents_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from the previous page's ``next_cursor``.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from the previous page's ``next_cursor``."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size, 1-100.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size, 1-100."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/chunks/{chunk_id}/revisions": {
      "post": {
        "tags": [
          "knowledge-revisions"
        ],
        "summary": "Create a pending chunk revision",
        "operationId": "create_chunk_revision_api_v1_documents__document_id__chunks__chunk_id__revisions_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "If-Match"
            }
          },
          {
            "name": "X-Actor-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "X-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkEditRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Chunk Revision Api V1 Documents  Document Id  Chunks  Chunk Id  Revisions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "knowledge-revisions"
        ],
        "summary": "List chunk revision history",
        "operationId": "list_chunk_revision_history_api_v1_documents__document_id__chunks__chunk_id__revisions_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Chunk Revision History Api V1 Documents  Document Id  Chunks  Chunk Id  Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/chunks/{chunk_id}/revisions/{revision_id}/rebuild": {
      "post": {
        "tags": [
          "knowledge-revisions"
        ],
        "summary": "Rebuild indexes and activate a pending revision",
        "operationId": "rebuild_chunk_revision_api_v1_documents__document_id__chunks__chunk_id__revisions__revision_id__rebuild_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rebuild Chunk Revision Api V1 Documents  Document Id  Chunks  Chunk Id  Revisions  Revision Id  Rebuild Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/chunks/{chunk_id}/revisions/{target_revision_id}/rollback": {
      "post": {
        "tags": [
          "knowledge-revisions"
        ],
        "summary": "Create a pending rollback revision",
        "operationId": "rollback_chunk_revision_api_v1_documents__document_id__chunks__chunk_id__revisions__target_revision_id__rollback_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          },
          {
            "name": "target_revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Target Revision Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "If-Match"
            }
          },
          {
            "name": "X-Actor-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "X-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkRollbackRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rollback Chunk Revision Api V1 Documents  Document Id  Chunks  Chunk Id  Revisions  Target Revision Id  Rollback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkEditRequest": {
        "properties": {
          "base_revision_id": {
            "type": "string",
            "minLength": 1,
            "title": "Base Revision Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "metadata_patch": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Patch"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "base_revision_id",
          "reason"
        ],
        "title": "ChunkEditRequest"
      },
...
      "ChunkRollbackRequest": {
        "properties": {
          "base_revision_id": {
            "type": "string",
            "minLength": 1,
            "title": "Base Revision Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "base_revision_id",
          "reason"
        ],
        "title": "ChunkRollbackRequest"
      },
...
      "DocumentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "type": "array",
            "title": "Items",
            "description": "Page elements."
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination cursor info."
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "DocumentListResponse",
        "description": "``GET /collections/{id}/documents`` response.\n\nExample (empty collection)::\n\n    {\n      \"items\": [],\n      \"page_info\": {\"next_cursor\": null, \"has_more\": false}\n    }",
        "examples": [
          {
            "items": [],
            "page_info": {
              "has_more": false
            }
          },
          {
            "items": [
              "<one element>"
            ],
            "page_info": {
              "has_more": true,
              "next_cursor": "eyJwYWdlIjoxfQ=="
            }
          }
        ]
      },
      "DocumentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "collection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Collection Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename."
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "description": "Raw file size in bytes."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "ready",
              "failed",
              "deleting"
            ],
            "title": "Status"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "description": "Number of chunks written to the vector store.",
            "default": 0
          },
          "image_count": {
            "type": "integer",
            "title": "Image Count",
            "description": "Number of images extracted and indexed.",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DocumentTagRef"
            },
            "type": "array",
            "title": "Tags",
            "description": "Tags bound to this document (B2.5 enrichment)."
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id",
            "description": "Folder this document is placed in, or null for the collection root (B2.5)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "collection_id",
          "filename",
          "size_bytes",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentSummary",
        "description": "Lightweight row for collection document list.",
        "example": {
          "chunk_count": 142,
          "collection_id": "b2c1f0e8-1234-5678-9abc-def012345678",
          "created_at": "2026-07-31T08:23:11.234Z",
          "filename": "employee-handbook.pdf",
          "id": "7d4a1c3e-2b0a-4d2c-8e9f-1a2b3c4d5e6f",
          "image_count": 7,
          "size_bytes": 1845632,
          "status": "ready",
          "updated_at": "2026-07-31T08:25:42.001Z"
        }
      },
      "DocumentTagRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Stable tag id (UUID string)."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Tag display name."
          },
          "color": {
            "type": "string",
            "title": "Color",
            "description": "Controlled colour token."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "DocumentTagRef",
        "description": "Tag summary embedded in a document list row (B2.5)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageInfo": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor for the next page. ``null`` means no more pages."
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Convenience flag; equivalent to ``next_cursor is not None``."
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "has_more"
        ],
        "title": "PageInfo",
        "description": "Cursor pagination metadata.\n\nThe cursor is opaque to clients; they should pass back whatever\n``next_cursor`` they got in the previous response.\n\n``next_cursor`` is required + nullable: every response carries the\nfield, with ``null`` meaning \"no more pages\". This avoids the\nfront-end having to distinguish between \"field missing\" and\n\"field is null\".",
        "example": {
          "has_more": true,
          "next_cursor": "eyJwYWdlIjoyfQ=="
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}